2020/05/24  add some Object

+++ music-server/src/main/java/com/example/demo/dao/ProductMapper.java
@@ -61,11 +61,12 @@ public interface ProductMapper {
      */
     int updateByPrimaryKey(Product record);
     
+    
     List<Product> allProducts();
 
     List<Product> productOfProductId(Integer productId);
 
-    List<Product> productOfId(Integer id);
+    List<Product> productOfId(Long id);
 
     List<Product> productOfProductName(String name);
 }
